Support Double shares for cryptos and Int shares for stocks, display today's change % next to portfolios in the widget, and add options to remove/rename portfolios

+++ Sources/StockTrackerCore/Models/Position.swift
@@ -5,10 +5,10 @@ import SwiftData
 public final class Position {
     @Attribute(.unique) public var id: UUID
     public var ticker: String
-    public var shares: Int
+    public var shares: Double
     public var portfolio: Portfolio?
     
-    public init(id: UUID = UUID(), ticker: String, shares: Int) {
+    public init(id: UUID = UUID(), ticker: String, shares: Double) {
         self.id = id
         self.ticker = ticker.uppercased().trimmingCharacters(in: .whitespacesAndNewlines)
         self.shares = shares
